feat: StaticContentController 적용
