+++ src/main/java/webserver/RequestHandler.java
@@ -1,7 +1,6 @@
 package webserver;
 
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Sets;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.util.StringUtils;
@@ -11,14 +10,14 @@ import utils.TemplateUtils;
 import java.io.*;
 import java.net.Socket;
 import java.net.URISyntaxException;
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
-    private static final String TEMPLATE_PATH = "./templates";
-    private static final String STATIC_PATH = "./static";
 
     private final Socket connection;
 
@@ -48,15 +47,17 @@ public class RequestHandler implements Runnable {
     }
 
     private Response handleRequest(HttpRequest httpRequest) throws Exception {
-        Controller controller = new RequestMapping(ImmutableMap.of(
+        return Optional.ofNullable(getController(httpRequest))
+                .orElseGet(StaticContentController::new)
+                .execute(httpRequest);
+    }
+
+    private Controller getController(HttpRequest httpRequest) {
+        return new RequestMapping(ImmutableMap.of(
                 "/usr/create", httpRequest1 -> new CreateUserController().execute(httpRequest1),
                 "/user/login", httpRequest2 -> new LoginController().execute(httpRequest2),
                 "/user/list", httpRequest3 -> new UserListController().execute(httpRequest3)))
                 .getController(httpRequest.getRequestURI());
-        if (controller == null) {
-            return null; // TODO 향후 static controller 로 교체한다
-        }
-        return controller.execute(httpRequest);
     }
 
     private void response(DataOutputStream dos, Response response) throws IOException {
@@ -64,18 +65,25 @@ public class RequestHandler implements Runnable {
             response302Header(dos, response.getHeaders());
         }
 
+        byte[] body = null;
         if (StringUtils.hasText(response.getViewName())) {
-            byte[] body = TemplateUtils.getTemplate(response.getViewName())
+            body = TemplateUtils.getTemplate(response.getViewName())
                     .apply(response.getModel())
                     .getBytes(UTF_8);
+        }
 
-            response200Header(dos, body.length);
+        if (response.getBody() != null) {
+            body = response.getBody();
+        }
+
+        if (body != null) {
+            response200Header(dos, body.length, response.getHeaders());
             responseBody(dos, body);
         }
     }
 
     private void responseStaticContent(String requestURI, DataOutputStream dos) throws IOException, URISyntaxException {
-        byte[] body = FileIoUtils.loadFileFromClasspath(getBasePath(requestURI) + requestURI);
+        byte[] body = FileIoUtils.loadFileFromClasspath(StaticContentController.getBasePath(requestURI) + requestURI);
         response200Header(dos, body.length, getContentType(requestURI));
         responseBody(dos, body);
     }
@@ -91,46 +99,40 @@ public class RequestHandler implements Runnable {
         }
     }
 
-    public static String getBasePath(String requestURI) {
-        return Sets.newHashSet("/css", "/js", "/fonts", "/images")
-                .stream()
-                .filter(requestURI::startsWith)
-                .findAny()
-                .map(path -> STATIC_PATH)
-                .orElse(TEMPLATE_PATH);
-    }
-
     private HttpRequest parseRequest(InputStream in) throws IOException {
         BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(in, UTF_8));
         RequestParser requestParser = new RequestParser(bufferedReader);
         return requestParser.parse();
     }
 
-    private void response200Header(DataOutputStream dos, int length) {
-        response200Header(dos, length, "text/html;charset=utf-8");
+    private void response200Header(DataOutputStream dos, int lengthOfBodyContent, String contentType) {
+        response200Header(dos, lengthOfBodyContent, Collections.singletonList("Content: " + contentType));
     }
 
-    private void response200Header(DataOutputStream dos, int lengthOfBodyContent, String contentType) {
+    private void response200Header(DataOutputStream dos, int length, List<String> headers) {
         try {
             dos.writeBytes("HTTP/1.1 200 OK \r\n");
-            dos.writeBytes("Content-Type: " + contentType + "\r\n");
-            dos.writeBytes("Content-Length: " + lengthOfBodyContent + "\r\n");
+            writeHeaders(dos, headers);
+            dos.writeBytes("Content-Length: " + length + "\r\n");
             dos.writeBytes("\r\n");
         } catch (IOException e) {
             logger.error(e.getMessage());
         }
     }
 
-
     private void response302Header(DataOutputStream dos, List<String> headers) {
         try {
             dos.writeBytes("HTTP/1.1 302 Found \r\n");
-            for (String header : headers) {
-                dos.writeBytes(header + "\r\n");
-            }
+            writeHeaders(dos, headers);
             dos.writeBytes("\r\n");
         } catch (IOException e) {
             logger.error(e.getMessage());
+        }
+    }
+
+    private void writeHeaders(DataOutputStream dos, List<String> headers) throws IOException {
+        for (String header : headers) {
+            dos.writeBytes(header + "\r\n");
         }
     }
